requirements.txt
pandas
openai
imaplib
email
configparser
python-dotenv